Doesn't include your commit today
I tried to add my work today to your commit but it is not working when I did so. So I've left yours asis and added this in a new folder so I don't mess up yours.
Added, InstallScreen with buttons to close app, and a button to close window and open new SplashScreen. I thought this would work temporarily so can check all windows.